Split up the fdtd, moving much of it to the helper.py
